address minor comments
